Update for CUDA 10
This adds support for the CUDA/Vulkan interop API, which requires
very new driver versions. As such, we allow the symbols to be
optionally loaded. I expect that when the next Video SDK release
comes out, it will also require the very new drivers and we'll make
these symbols compulsory.