Fix CUdeviceptr definition for 64bit CPU
NVIDIA driver implementation defines and expects ‘CUdeviceptr’ to
be 64-bit wide on the 64-bit platforms.
FFMPEG build on PPC64 architecture on the other hand, falls back to
an incorrect (32-bit wide data type) definition of this data type.
This patch adds a new build time check to pick up the correct definition
on PPC64 architecture.

Signed-off-by: BtbN <btbn@btbn.de>